Add lifecycle management to entities
Add a way for entities to be removed when they go off the screen. This
still has entities maintain their encapsulation with them handling when
they are off the screen, but we also need to notify any parents
(composites) as these objects will still have a reference to the objects
that are no-longer needed.

It feels a bit over-engineered, I guess I'll find out later.