Dark theme, strobe event block and fix bugs
Default to darker colours

Strobe the event block when a ball is fired out. Balls now take a colour
based on the EventBlockElement (which will probably in turn take it
from the EventBlock.

Don't emit events when the window isn't visible. This causes balls to be
accumulated and "blow up" when the window is visible again.